requirements.txt
BetterConfigAJM==0.1
lxml==5.3.0
pillow==10.4.0
python-docx==1.1.2
typing_extensions==4.12.2
